requirements.txt
# Core Application Dependencies
fastapi
uvicorn[standard]  # Installs uvicorn with necessary dependencies
pydantic
requests

# LangGraph & LangChain RAG Pipeline
langchain
langgraph
langchain-openai
langchain-community

# Document Processing & Vector Stores
PyPDF2
PyMuPDF
tabulate
pypdf # For PDF handling
python-docx
pinecone
langchain-pinecone

# Hybrid Search & Reranking
rank_bm25  # For BM25 keyword search
sentence-transformers # For cross-encoder reranking

# Core Libraries
cohere
openai
tiktoken # Used by OpenAI models for token counting

# Utility & Deployment (Optional but recommended)
python-dotenv
gunicorn